+++ util-stats/src/main/scala/com/twitter/finagle/stats/NameTranslatingStatsReceiver.scala
@@ -20,18 +20,19 @@ abstract class NameTranslatingStatsReceiver(
   override def counter(counterSchema: CounterSchema): Counter = {
     self.counter(
       CounterSchema(
-        counterSchema.metricBuilder.withName(translate(counterSchema.metricBuilder.name))))
+        counterSchema.metricBuilder.withName(translate(counterSchema.metricBuilder.name): _*)))
   }
 
   override def stat(histogramSchema: HistogramSchema): Stat = {
     self.stat(
       HistogramSchema(
-        histogramSchema.metricBuilder.withName(translate(histogramSchema.metricBuilder.name))))
+        histogramSchema.metricBuilder.withName(translate(histogramSchema.metricBuilder.name): _*)))
   }
 
   override def addGauge(gaugeSchema: GaugeSchema)(f: => Float): Gauge = {
     self.addGauge(
-      GaugeSchema(gaugeSchema.metricBuilder.withName(translate(gaugeSchema.metricBuilder.name))))(f)
+      GaugeSchema(
+        gaugeSchema.metricBuilder.withName(translate(gaugeSchema.metricBuilder.name): _*)))(f)
   }
 
   override def toString: String = s"$self/$namespacePrefix"